add URL checking in tr_add, small fixes
Forgot to move positional parmeters around when I moved func to a
variable in trop_awk.

For startup, I added a check to see if the user issuing -startup is
already logged in; If they are, don't do anything. Since -startup
only needs to be used once, this stops it from running everytime
you log into a vtty or SSH into your machine.

Fixed up tracker_add() in trop.awk, in the tr_match_other call I
forgot to swap the parameters around, and I added a function for
URL checking.