fix some bugs from last commit
Description of last commit (439915e):
=====================================
439915e introduced torrent specific functions to -ns, -dl and -m.

It also added the -version switch.

It moved alot of stuff around, attempting to remove some code
duplication with the introduction of the new functions and trop_private
calls.
=====================================

This commit fixed the common test for the tracker-alias and trop_private;
A second case was need to check through the common flags.

Patched up trop.{awk,sh} a bit, fixing various bugs. There are likely
still a few regressions in the code, however.