fix trop_torrent_done() conditional
The code assumed that every torrent in the queue would be done and
would die if a torrent wasn't finished, causing torrents to not be
removed on time and unintuitive log messages.

Also, I screwed up the versioning so I have to go back and rewrite
all the commits. Please reset --hard all branches.