disallow `#' as the first char of an alias
What I stated in commit f926b26 was only true if the whole alias
was just one character. If it were two or more characters, it
would not match the regex. I incorrectly assumed trop.awk didn't
match the EOL in the regex.

However, to avoid any confusion I think disallowing `+' or `#' as
the starting character would be the better choice (`#' represents
a comment).