connect trop_awk tth call with pipe
Because the ARGV array is removed at the beginning of trop.awk,
all files that were supplied with the call to awk needs to be
referenced from a variable that stores the file name. In
tr_total_hashop I originally called the `alias`_thash file implicitly
in the do while getline loop. Solution: store the hashfile in a
variable and getline the file directly.