add install.sh update check, small touchups
install.sh update now checks for user modifiable file changes by
storing a vanilla checksum of each file in the trop cache directory.
Of course, this will only be useful after _another_ update of one of
these files as it needs something to compare the current files to.

Other changes:
- trop.sh: Moved some functions into a more appropriate location.
           Apparently an implicit `for i; do ...' requires the `do'
           keyword to be placed on a new line in some sh implementations.
- trop.conf: Decided that if a user has anything other than a valid
             `yes' as a value for an option then it will default to
             false, functionally.