show total upload amount, add flag suggestions
The sul flag now shows the total UL rates of seeding torrents.
This was implemented differently than it's tracker equivalent
(will be) because tr-remote outputs the sum itself; So instead
of calculating it in AWK I just grabbed it from tr-remote.

trop will now display some flag suggestions if the user enters
an unrecognized flag, like git does except for now the rules
for which flags to display are probably simpler in trop's case.

Other changes
trop.sh:
 - Some style changes.

trop.awk:
 - Change variable `ll' to `longest_name'. This was missed from a
   previous commit.
 - Added a function to convert kB/s output to different magnitudes.