rework -sul
It is much faster to get sul info from a tr-remote -l invocation.
Because we don't need to worry about anything tracker related,
using trop_seed_info is a bit overkill. This also fixes a syncing
problem with the total output, because the UL rate will likely have
changed between the storing of the Sum line and each seed_info call.