trop.awk: -tdl: add break when state doesn't match
In tracker_dl_info() when the tracker_match_line() if blk was
entered, the while (getline) loop would iterate until it found a
torrent that is in a downloading state; This caused the name and
ID lines to be incorrect if tracker_match_line() matched a seeding
torrent that wasn't in a downloading state before the successive
downloading torrent _was_ matched. The solution was to always break
out if the State line was matched.